requirements.txt
asn1crypto==1.5.1
asttokens==2.0.5
awswrangler==2.14.0
backcall==0.2.0
beautifulsoup4==4.10.0
boto3==1.21.24
botocore==1.24.24
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.31.2
greenlet==1.1.2
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
jmespath==1.0.0
jsonpath-ng==1.5.3
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
lxml==4.8.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
nest-asyncio==1.5.4
netadata @ file:///C:/neta-data/neta-data
numpy==1.22.3
openpyxl==3.0.9
opensearch-py==1.1.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pg8000==1.22.1
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
progressbar2==3.55.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==6.0.1
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-utils==3.1.0
pytz==2021.3
pywin32==303
pyzmq==22.3.0
redshift-connector==2.0.905
requests==2.27.1
requests-aws4auth==1.1.1
s3transfer==0.5.2
scipy==1.8.0
scramp==1.4.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
varietybenchmark @ file:///C:/neta-data/variety-benchmark
wcwidth==0.2.5
